refactor: preview, select persona등 로직 분리
