+++ apps/web/src/app/[locale]/mypage/(github-custom)/LinePersonaSelect.tsx
@@ -1,0 +1,60 @@
+'use client';
+
+import React, { useState } from 'react';
+import { useTranslations } from 'next-intl';
+import { css } from '_panda/css';
+import { Button } from '@gitanimals/ui-panda';
+
+import { SelectPersonaList } from '../PersonaList';
+
+export const LinePersonaSelect = ({
+  selectPersona,
+  onChangePersona,
+}: {
+  selectPersona: string | null;
+  onChangePersona: (personaId: string) => void;
+}) => {
+  const t = useTranslations('Mypage');
+
+  const [isExtend, setIsExtend] = useState(false);
+
+  return (
+    <>
+      <section className={selectPetContainerStyle}>
+        <h2 className="heading">{t('change-pet')}</h2>
+        <Button className="extend-button" onClick={() => setIsExtend((prev) => !prev)}>
+          {isExtend ? t('shrink-button') : t('extend-button')}
+        </Button>
+      </section>
+      <section className={isExtend ? flexGrowSectionStyle : ''}>
+        <SelectPersonaList
+          selectPersona={selectPersona ? [selectPersona] : []}
+          onSelectPersona={(persona) => onChangePersona(persona.id)}
+          isExtend={isExtend}
+        />
+      </section>
+    </>
+  );
+};
+
+const flexGrowSectionStyle = css({
+  flex: '1',
+  minHeight: '0',
+  overflow: 'auto',
+  display: 'flex',
+  flexWrap: 'wrap',
+});
+
+const selectPetContainerStyle = css({
+  position: 'relative',
+  '& .heading': {
+    textStyle: 'glyph18.bold',
+    color: 'white',
+    marginBottom: '16px',
+  },
+  '& .extend-button': {
+    position: 'absolute',
+    top: '-16px',
+    right: 0,
+  },
+});
